Rename modular::builder module to modular::parser
Better reflects what the module contains.
